Rename lita-slack to shopify-lita-slack
