+++ lib/shopify-lita-slack.rb
